deps: move changed files
